+++ include/heuristics.h
@@ -4,6 +4,6 @@
 #include <cmath>
 #include "types.h"
 
-double heuristicDiagonalDistance(Location state, Location target);
+uint32_t heuristicDiagonalDistance(Location state, Location target);
 
 #endif //LABYRINTHROBOTS_HEURISTICS_H
